moved json cleaner into streams-processor-jackson better javadoc comment
